requirements.txt
httpx==0.27.0
ipdb==0.13.13
openai==1.37.1
pandas==2.2.2
pyautogen==0.2.28
PyYAML==6.0.1
quart==0.18.4
quart_cors==0.7.0
replicate==0.27.0
Requests==2.32.3
colorama==0.4.6
werkzeug==2.3.7
